backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pillow==10.1.0
passlib==1.7.4
email-validator==2.1.1
tensorflow
psycopg2-binary